+++ CHANGES.md
@@ -118,6 +118,7 @@
  * added missing translation (echo 'blabla';)
  * All existing translation string are now in english (some were in french) Ready for translation (.pot refactored)
  * Add if ( !defined( 'ABSPATH' ) ) exit; on each php code (not templates)
+ * fontes Google peu utiles, souvent urchargées par Roboto supprimées: Merriweather Montserrat et Inconsolata
         
 ##### CSS changes  
  * page rubriques (par exemple /le-laboratoire/